+++ sb3_contrib/ppo_mask/__init__.py
@@ -1,0 +1,2 @@
+from sb3_contrib.ppo_mask.policies import CnnPolicy, MlpPolicy  # , MultiInputPolicy
+from sb3_contrib.ppo_mask.ppo_mask import MaskablePPO